ticket notification id, find user by name

+++ src/main/java/com/slmas/Sl/repository/TicketRepository.java
@@ -7,7 +7,7 @@ import java.util.Date;
 import java.util.List;
 
 public interface TicketRepository {
-    Integer createTicket (Ticket ticket) throws RepositoryException;
+    Long createTicket (Ticket ticket) throws RepositoryException;
     List<Ticket> getUserTickets(Date startDate, Date endDate, Long userId) throws RepositoryException;
     Ticket getTicketById (Long id) throws RepositoryException;
     List<Ticket> getFilteredTickets(Date startDate, Date endDate, String area, String closed) throws RepositoryException;
